--- main/components/Privacy.jsx
@@ -1,517 +1,0 @@
-import React from 'react'
-import styles from './Privacy.module.scss';
-
-function PrivacyPolicy() {
-    return (
-      <section style={{ paddingTop: '10px',backgroundColor:'#fff' }} className={styles.container}>
-        <p className='c1' >
-          <span className='c7 c4' >Digital Aided School </span>
-        </p>
-        <p className='c1'>
-          <span className='c4 c7'>Privacy Policy</span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Website</span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>What We Collect and Why</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c6'>
-            &nbsp;collects some anonymous information each time you visit its
-            website, so it can improve the overall quality of your online
-            experience. It collects your IP address, referral data, and browser
-            and platform type. You do not have to register with{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;before we can collect this anonymous information.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c6'>
-            website does not require you to share information that identifies you
-            personally, such as your name or e-mail address, for you to use the
-            website. The{' '}
-          </span>
-          <span className='c4'>Digital Aided School &nbsp;</span>
-          <span className='c6'>
-            website may assign an anonymous ID number to your request(s) and links
-            the following additional data to that number: the date and time you
-            visited the website, your search terms, and the links upon which you
-            choose to click. Like most standard website servers, we use log files
-            to collect and store this anonymous user information.{' '}
-          </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            &nbsp;analyzes the information to examine trends, administer the site,
-            track user&rsquo;s movement in the aggregate, and gather broad
-            demographic information for aggregate use.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>From time-to-time the </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>
-            &nbsp;website requests information from you that may identify you
-            personally in addition to the anonymous information described above.
-            The requested information typically includes contact information (such
-            as name, e-mail address, and zip code), and demographic information
-            (such as age, gender, occupation, and household income).{' '}
-          </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            &nbsp;will not use or disclose your individual user information except
-            as described in this privacy policy and without first obtaining your
-            consent.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Cookies</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>&nbsp;may set and access </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            &nbsp;or partner cookies on your computer. A cookie is a small data
-            file that a website may send to your browser and which may then be
-            stored on your system. We may occasionally use both session cookies
-            and persistent cookies. For the session cookie, once you close the
-            browser, the cookie simply terminates. A persistent cookie is a small
-            text file stored on your hard drive for an extended period of time.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Persistent cookies can be removed by following Internet browser help
-            file directions.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            We use cookie and web beacon technology to track which pages on the{' '}
-          </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c6'>
-            &nbsp;websites our visitors view and which Web browsers our visitors
-            use. We may also use cookies to help us provide personalized features
-            and preferences to you within various{' '}
-          </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            &nbsp;websites in a way that may be linked to individual user
-            information that you elect to share with us. We may also use
-            &ldquo;action tags&rdquo; to assist in delivering the cookie. These
-            action tags allow us to compile statistics and other data about our
-            customers on an aggregate basis, by embedding a random identifying
-            number on each user&rsquo;s browser. Then we track where that number
-            shows up on our site.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Special Notification with respect to Children&rsquo;s Privacy (Users
-            under the age of 13)
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            In response to concerns about protecting children&rsquo;s privacy
-            online, Congress enacted the Children&rsquo;s Online Privacy
-            Protection Act of 1998 (&ldquo;COPPA&rdquo;), which sets forth rules
-            and procedures governing the ways in which websites may collect, use
-            and disclose any personal information for children under the age of
-            13. In accordance with{' '}
-          </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            &nbsp;policy and COPPA regulations, we DO NOT:
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Request or knowingly collect personally identifiable information
-            online or offline contact information from users under 13 years of
-            age; or knowingly use or share personal information from users under
-            13 years of age with third parties.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            It is possible that by fraud or deception we may receive information
-            given to us or pertaining to children under 13. If we are notified of
-            this, as soon as we verify the information, we will immediately delete
-            the information from our servers.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Choice/Opt-out</span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            We desire to keep you in control of the personal information you
-            provide to us. Accordingly, you can review, correct, change or remove
-            the personal registration information you provide to{' '}
-          </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c6'>&nbsp;and that </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>&nbsp;controls.</span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            If you do not want to receive e-mail from us in the future, please let
-            us know by sending us email at hello@digitalaidedschool.com
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            If you supply us with your postal address online, you may receive
-            periodic mailings from us with information on new products and
-            services or upcoming events. If you do not wish to receive such
-            mailings, please let us know by sending us email at
-            hello@digitalaidedschool.com
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            Persons who supply us with their telephone numbers online will only
-            receive telephone contact from us with information regarding applying
-            to the{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            . If you would not like to be contacted, please provide us with your
-            name and phone number to ensure that we remove the correct person from
-            the list.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Sharing</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>
-            &nbsp;will make every effort to preserve your privacy.{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>
-            &nbsp;may need to disclose personal information when required by law
-            wherein we have a good-faith belief that such action is necessary to
-            comply with a current judicial proceeding, a criminal investigation, a
-            court order, or legal process served on our website. We may also share
-            your personal information if we believe it is necessary in order to
-            investigate, prevent, or take action regarding illegal activities,
-            suspected fraud, situations involving potential threats to the
-            physical safety of any person, violations of{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>&nbsp;or such </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;websites&rsquo; terms of use, or as otherwise required by law.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Legal Notice</span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            Notwithstanding any language to the contrary, nothing contained herein
-            constitutes nor is intended to constitute an offer, inducement,
-            promise, or contract of any kind. The data contained herein is for
-            informational purposes only and is not represented to be error free.
-            Any links to information not owned by{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>
-            &nbsp; provided as a courtesy. They are not intended to, nor do they
-            constitute an endorsement by{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;of the linked materials, products, services, or information
-            contained therein.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;do not have any control over the privacy practices of these
-            websites or information, and all visitors of these websites should
-            read and understand the policies and privacy practices for external
-            sites.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Copyright Infringement</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c6'>
-            &nbsp;is committed to complying with INDIAN copyright law and
-            respecting the rights of copyright holders, while simultaneously
-            defending the academic freedom of faculty and providing the optimal
-            learning experience for students. Faculty and staff will make a good
-            faith effort to only use or distribute third-party copyrighted
-            materials as allowed by purchase, license, fair use, or other legal
-            provision. Anyone who participates in the unauthorized distribution of
-            copyrighted material may be subject to civil and criminal liabilities,
-            in addition to sanctions imposed by{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>. </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Lawful Basis for Collecting and Processing of Personal Data
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            In order for educational institutions to educate their students both
-            in className and on-line, it is essential and necessary that they have
-            lawful bases to collect, process, use, and maintain data for its
-            student, employees, applicants, and others involved in its educational
-            programs. The lawful bases include, without limitation, admission,
-            registration, delivery of classNameroom, on-line, and study abroad
-            education, grades, communications, employment, program analysis for
-            improvements, and records retention. Examples of data that{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;and/or its affiliates may need to collect in connection with the
-            lawful bases are name, email address, phone number, IP address,
-            physical address or other location identifier, photos, as well as some
-            sensitive personal data obtained with prior consent.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Most of the collection and processing of personal data will include:
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            Processing necessary for the purposes of the legitimate interests
-            pursued by{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;in providing educational opportunities and employment.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Processing necessary for the performance of a contract to which the
-            subject is party or in order to take steps at the request of the data
-            subject prior to entering into a contract.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            Processing necessary for compliance with a legal obligation to which{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>&nbsp; is subject.</span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            The data subject has given consent to the processing of his or her
-            personal data for one or more specific purposes. This lawful basis
-            pertains primarily but not exclusively to the protection of research
-            subjects.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            There will be some instances where the collection and processing of
-            personal data will be pursuant to other lawful bases.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Types of Personal Data Collected and Why</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;collect a variety of personal and sensitive data to meet one of
-            its lawful bases, as referenced above. Most often the data is used for
-            admission, registration, delivery of classNameroom, on-line, and study
-            abroad education, grades, communications, employment, program analysis
-            for improvements, and records retention. If you have specific
-            questions regarding the collection and use of your personal data.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            If a data subject refuses to provide personal data that is required,
-            such refusal may make it impossible for{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;to provide education, employment, research or other requested
-            services.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Personal data typically includes, but is not limited to:
-          </span>
-        </p>
-        <a id='t.136061b4cf4bc2dc21e007a429faab75e5a66403'></a>
-        <a id='t.0'></a>
-        <table className='c13'>
-          <tbody>
-            <tr className='c2'>
-              <td className='c11' colSpan='1' rowSpan='1'>
-                <p className='c5'>
-                  <span className='c8 c3'>Data Type</span>
-                </p>
-              </td>
-              <td className='c9' colSpan='1' rowSpan='1'>
-                <p className='c5'>
-                  <span className='c3 c8'>Source</span>
-                </p>
-              </td>
-            </tr>
-            <tr className='c12'>
-              <td className='c11' colSpan='1' rowSpan='1'>
-                <p className='c5'>
-                  <span className='c8 c3'>
-                    First name, last name, address, and any other contact
-                    information that can directly or indirectly identify the data
-                    subject.
-                  </span>
-                </p>
-              </td>
-              <td className='c9' colSpan='1' rowSpan='1'>
-                <p className='c5'>
-                  <span className='c8 c3'>
-                    Collected directly from you, the data subject, through:
-                    application forms, online forms, emails, phone calls, chats,
-                    SMS, shared files.
-                  </span>
-                </p>
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        <p className='c1'>
-          <span className='c0'>
-            Individual Rights of the Data Subject under the EU GDPR
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>Individual data subjects covered by </span>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            Privacy Policy will be afforded the following rights:
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right of access &ndash; you have the right to request a copy of the
-            information that we hold about you.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right of rectification &ndash; you have a right to correct data that
-            we hold about you that is inaccurate or incomplete.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right to be forgotten &ndash; in certain circumstances you can ask for
-            the data we hold about you to be erased from our records.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right to restriction of processing &ndash; where certain conditions
-            apply to have a right to restrict the processing.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right of portability &ndash; you have the right to have the data we
-            hold about you transferred to another organization.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right to object &ndash; you have the right to object to certain types
-            of processing such as direct marketing.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>
-            Right to object to automated processing, including profiling &ndash;
-            you also have the right to be subject to the legal effects of
-            automated processing or profiling.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>Right to judicial review &ndash;</span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;refuses your request under rights of access, we will provide you
-            with a reason as to why.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            Right to complain &ndash; if you wish to make a complaint about how
-            your personal data is being processed by{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-        </p>
-        <p className='c1'>
-          <span className='c6'>
-            All personal data and sensitive personal data collected or processed
-            by{' '}
-          </span>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;under the scope of the Privacy Policy must comply with the
-            security controls and safeguards of FERPA, GLBA, and PCI as set forth
-            in the Privacy Policy.
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Retention Period</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School </span>
-          <span className='c0'>
-            and/or its affiliates will process and store personal data for no
-            longer than it is necessary for the specified above purpose(s).
-          </span>
-        </p>
-        <p className='c1'>
-          <span className='c0'>Policy Updates</span>
-        </p>
-        <p className='c1'>
-          <span className='c4'>Digital Aided School</span>
-          <span className='c0'>
-            &nbsp;reserves the right to modify this privacy policy at any time. We
-            encourage visitors to frequently check this page for any changes to
-            this Privacy Policy. If we make changes, we will post an updated
-            effective date below. Your continued use of this site after any change
-            in this Privacy Policy will constitute your acceptance of such change.
-            This policy has been updated on Mar 18, 2022.
-          </span>
-        </p>
-        <p className='c5 c10'>
-          <span className='c7 c15'></span>
-        </p>
-      </section>
-    )
-  }
-  
-  export default PrivacyPolicy
-
