--- main/components/Privacy.module.scss
@@ -1,5 +1,0 @@
-
-
-.container{
-    background-color:white
-}
